+++ src/views/secret/SecretEdition.vue
@@ -13,27 +13,68 @@
                 </router-link>
             </div>
         </div>
-        <form class="top-space" @submit.prevent="submit()">
+        <form class="top-space" @submit.prevent="confirm()">
             <div class="field">
                 <label class="label">Topic</label>
                 <div class="control">
                     <input class="input" type="text" placeholder="Secret topic ..." v-model="secret.topic" required>
                 </div>
             </div>
-            <div class="field">
-                <label class="label">Username/Email</label>
-                <div class="control">
-                    <input class="input" type="text" placeholder="Username ..." v-model="secret.username" required>
+            <label class="label">Username/Email</label>
+            <div class="field is-grouped">
+                <div class="control is-expanded">
+                    <input 
+                        class="input" 
+                        :type="usernameVisibility ? 'text' : 'password'" 
+                        placeholder="Username ..." 
+                        v-model="secret.username"
+                        required
+                    >
                 </div>
-                <SecretGenerator @secret="usernameGenerated"></SecretGenerator>
-            </div>
-            <div class="field">
-                <label class="label">Secret/Password</label>
                 <div class="control">
-                    <input class="input" type="text" placeholder="Secret ..." v-model="secret.secret">
+                    <button type="button" class="button is-dark" @click="toggleVisibility('username')">
+                        <div class="icon">
+                            <i class="fa-solid fa-lg fa-eye" v-if="!usernameVisibility"></i>
+                            <i class="fa-solid fa-lg fa-eye-slash" v-if="usernameVisibility"></i>
+                        </div>
+                    </button>
                 </div>
-                <SecretGenerator @secret="secretGenerated"></SecretGenerator>
+                <div class="control">
+                    <button type="button" class="button is-dark" @click="copy('username')">
+                        <div class="icon">
+                            <i class="fa-solid fa-lg fa-copy"></i>
+                        </div>
+                    </button>
+                </div>
             </div>
+            <SecretGenerator @secret="usernameGenerated"></SecretGenerator>
+            <label class="label">Secret/Password</label>
+            <div class="field is-grouped">
+                <div class="control is-expanded">
+                    <input 
+                        class="input" 
+                        :type="passwordVisibility ? 'text' : 'password'" 
+                        placeholder="Secret ..." 
+                        v-model="secret.secret"
+                    >
+                </div>
+                <div class="control">
+                    <button type="button" class="button is-dark" @click="toggleVisibility('secret')">
+                        <div class="icon">
+                            <i class="fa-solid fa-lg fa-eye" v-if="!passwordVisibility"></i>
+                            <i class="fa-solid fa-lg fa-eye-slash" v-if="passwordVisibility"></i>
+                        </div>
+                    </button>
+                </div>
+                <div class="control">
+                    <button type="button" class="button is-dark" @click="copy('secret')">
+                        <div class="icon">
+                            <i class="fa-solid fa-lg fa-copy"></i>
+                        </div>
+                    </button>
+                </div>
+            </div>
+            <SecretGenerator @secret="secretGenerated"></SecretGenerator>
             <div class="field is-grouped">
                 <div class="control">
                     <button class="button" 
@@ -67,6 +108,7 @@
                 </div>
             </div>
         </form>
+        <Modal :state="modal" @submit="submit" @close="toggleModal(false, '', '')"></Modal>
     </div>
 </template>
 
@@ -74,6 +116,7 @@
 import type { Secret } from '@/models/secret';
 import { SecretService } from '@/services/secret-service';
 import SecretGenerator from '@/components/SecretGenerator.vue';
+import Modal from '@/components/Modal.vue';
 
 export default {
     created() {
@@ -110,28 +153,55 @@ export default {
                 subtitle: "",
                 edition: true
             },
-            secret: secret
+            secret: secret,
+            usernameVisibility: false,
+            passwordVisibility: false,
+            modal: {
+                active: "",
+                title: "",
+                button: ""
+            }
         };
     },
     methods: {
         deleteSecret() {
-            this.secretRepo.delete(this.$route.params.categoryId, this.secret.id);
-            this.$router.push("/categories/" + this.$route.params.categoryId);
+            this.toggleModal(true, "Delete secret", "Delete");
         },
-        submit() {
-            this.secret.username = this.secretService.encrypt(this.secret.username);
-            this.secret.secret = this.secretService.encrypt(this.secret.secret);
-            this.secretRepo.save(this.$route.params.categoryId, this.secret);
-            this.$router.push("/categories/" + this.$route.params.categoryId);
+        submit(type: string) {
+            if(type == "Delete") {
+                this.secretRepo.delete(this.$route.params.categoryId, this.secret.id);
+                this.$router.push("/categories/" + this.$route.params.categoryId);
+            } else {
+                this.secret.username = this.secretService.encrypt(this.secret.username);
+                this.secret.secret = this.secretService.encrypt(this.secret.secret);
+                this.secretRepo.save(this.$route.params.categoryId, this.secret);
+                this.$router.push("/categories/" + this.$route.params.categoryId);
+            }
+        },
+        confirm() {
+            this.toggleModal(true, this.setting.edition ? "Update secret" : "New secret", this.setting.edition ? "Update" : "Add")
         },
         usernameGenerated(secret: string) {
             this.secret.username = secret
         },
         secretGenerated(secret: string) {
             this.secret.secret = secret
+        },
+        toggleVisibility(type: string) {
+            if(type == 'username') this.usernameVisibility = !this.usernameVisibility
+            else this.passwordVisibility = !this.passwordVisibility
+        },
+        copy(type: string) {
+            navigator.clipboard.writeText(type == 'username' ? this.secret.username : this.secret.secret);
+            console.log('copied successfully!')
+        },
+        toggleModal(active: boolean, title: string, button: string) {
+            this.modal.active = active ? "is-active" : "";
+            this.modal.title = title;
+            this.modal.button = button
         }
     },
-    components: { SecretGenerator }
+    components: { SecretGenerator, Modal }
 }
 </script>
